refactor: split bot into packages
